# src/public/My Project.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "My Project",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/bookmarks": {
      "get": {
        "summary": "북마크 리스트 조회",
        "deprecated": false,
        "description": "",
        "tags": ["travely-server/bookmark"],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "",
            "required": false,
            "example": "672901f5f3339eca4777fc11",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "success": true,
                      "data": {
                        "bookmarks": [
                          {
                            "id": "672905e83f52cfeef7356cc3",
                            "thumbnail": "https://i.imgur.com/kMVnKYx.jpeg",
                            "travelTitle": "부산 여행",
                            "tag": ["Nature"],
                            "bookmark": true,
                            "createdBy": {
                              "userId": "672901f5f3339eca4777fc11",
                              "userName": "nakyenko4"
                            },
                            "review": {
                              "travelScore": 4.5,
                              "reviewCnt": 2
                            },
                            "createdAt": "2024-11-04T17:35:36.014Z"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "북마크 삭제",
        "deprecated": false,
        "description": "",
        "tags": ["travely-server/bookmark"],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "",
            "required": false,
            "example": "672901f5f3339eca4777fc11",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "travelId",
            "in": "query",
            "description": "",
            "required": false,
            "example": "672905e83f52cfeef7356cc3",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "userId": "672901f5f3339eca4777fc11",
                "travelId": "672905e83f52cfeef7356cc3",
                "isBookmark": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "success": true,
                      "data": {
                        "id": "67a6dcf4b90693dcf3a10bde",
                        "userId": "672901f5f3339eca4777fc11",
                        "isBookmark": false
                      }
                    }
                  },
                  "2": {
                    "summary": "Success",
                    "value": {
                      "success": true,
                      "data": {
                        "id": "672905e83f52cfeef7356cc3",
                        "userId": "672901f5f3339eca4777fc11",
                        "isBookmark": true
                      }
                    }
                  },
                  "3": {
                    "summary": "Success",
                    "value": {
                      "success": true,
                      "data": {
                        "id": "672905e83f52cfeef7356cc3",
                        "userId": "672901f5f3339eca4777fc11",
                        "isBookmark": true
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "북마크 추가",
        "deprecated": false,
        "description": "",
        "tags": ["travely-server/bookmark"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "traveld": {
                    "type": "string"
                  }
                },
                "required": ["userId", "traveld"]
              },
              "example": {
                "userId": "672901f5f3339eca4777fc11",
                "travelId": "678718903d61d821c6f91ac5"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "success": true,
                      "data": {
                        "id": "67a6dcf4b90693dcf3a10bde",
                        "userId": "672901f5f3339eca4777fc11",
                        "isBookmark": true
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    }
  },
  "components": {
...
    "securitySchemes": {}
  },
  "servers": [
    {
      "url": "http://3.37.101.147:3000/api",
      "description": "travely-server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "travely-server-local"
    }
  ],
  "security": {}
}